+++ core/dl-meow.el
@@ -31,10 +31,24 @@ being drawn; lambda-line caches the truly active one in
                     'face 'dl-meow-indicator-inactive)
       s)))
 
+(defun dl-meow--magit-unbind-jk ()
+  "Free magit's `j' / `k' from meow normal-state shadowing."
+  (local-unset-key (kbd "j"))
+  (local-unset-key (kbd "k")))
+
 (use-package meow
+  :custom
+  (meow-use-cursor-position-hack t)
+  (meow-use-clipboard t)
+  (meow-goto-line-function 'consult-goto-line)
   :config
-  (setq meow-mode-state-list
-        (append meow-mode-state-list '((vterm-mode . insert))))
+  (setq meow--kbd-delete-char "<deletechar>")
+  (meow-thing-register 'angle '(regexp "<" ">") '(regexp "<" ">"))
+  (add-to-list 'meow-char-thing-table '(?a . angle))
+  (add-to-list 'meow-mode-state-list '(vterm-mode        . insert))
+  (add-to-list 'meow-mode-state-list '(magit-status-mode . normal))
+  (add-to-list 'meow-mode-state-list '(magit-log-mode    . normal))
+  (add-hook 'magit-mode-hook #'dl-meow--magit-unbind-jk)
   (meow-setup)
   (meow-global-mode 1)
   (dl-meow--apply-indicator-faces)