+++ editing/dl-multi-edit.el
@@ -11,15 +11,5 @@
 
 (use-package iedit)
 
-;; (setq-local combobulate-checkout-path (expand-file-name "checkout/combobulate" user-emacs-directory))
-;; (use-package combobulate
-;;   :vc (:url "https://github.com/mickeynp/combobulate.git")
-;;   :custom
-;;   ;; You can customize Combobulate's key prefix here.
-;;   ;; Note that you may have to restart Emacs for this to take effect!
-;;   (combobulate-key-prefix "C-c o")
-;;   :hook ((prog-mode . combobulate-mode)))
-
-
 (provide 'dl-multi-edit)
 ;;; dl-multi-edit.el ends here